chore: rename datasets feature to dataset
