+++ src/vectorize/dataset/exceptions.py
