+++ src/vectorize/dataset/utils/validate_zip.py
