+++ src/vectorize/synthesis/tasks.py
@@ -10,12 +10,12 @@ from sqlmodel.ext.asyncio.session import AsyncSession
 from vectorize.common.task_status import TaskStatus
 from vectorize.config import settings
 from vectorize.config.db import engine
-from vectorize.datasets.classification import Classification
-from vectorize.datasets.exceptions import DatasetNotFoundError
-from vectorize.datasets.models import Dataset
-from vectorize.datasets.repository import get_dataset_db, upload_dataset_db
-from vectorize.datasets.upload_options_model import DatasetUploadOptions
-from vectorize.datasets.utils.save_dataset import _save_dataframe_to_fs
+from vectorize.dataset.classification import Classification
+from vectorize.dataset.exceptions import DatasetNotFoundError
+from vectorize.dataset.models import Dataset
+from vectorize.dataset.repository import get_dataset_db, upload_dataset_db
+from vectorize.dataset.upload_options_model import DatasetUploadOptions
+from vectorize.dataset.utils.save_dataset import _save_dataframe_to_fs
 
 from .repository import update_synthesis_task_status
 from .text_extractor import extract_text_from_media
