+++ src/vectorize/utils/routers.py
@@ -5,7 +5,7 @@ from fastapi import APIRouter, FastAPI
 from vectorize.ai_model.router import router as models_router
 from vectorize.common.router import router as common_router
 from vectorize.config.config import settings
-from vectorize.datasets.router import router as dataset_router
+from vectorize.dataset.router import router as dataset_router
 from vectorize.inference.router import router as embeddings_router
 from vectorize.synthesis.router import router as synthesis_router
 from vectorize.upload.router import router as upload_router
